backend/requirements.txt
Django==4.2.7
celery==5.3.6
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg[binary]==3.2.6
python-dotenv==1.0.0
redis==5.0.1
